Add files via upload
Pokerito (it's almost Poker!)

Pokerito is like Poker, but the rules are a lot simpler.

The dealer will give you one card
and the computer one card.
Then, the dealer will draw five cards (the river).
The player with the most river matches wins. If the matches are equal, everyone's a winner!

Start the game by explaining the rules.

Let's play Pokerito. Type anything when you're ready.

It's like Poker, but a lot simpler.

- There are two players, you and the computer.

- The dealer will give each player one card.

- Then, the dealer will draw five cards (the river)

- The player with the most river matches wins!

- If the matches are equal, everyone's a winner!

- Ready? Type anything if you are.